refactor(graphql): move the getQueryResult utility functions to new files
LIIKUNTA-578
