+++ graphql/src/datasources/es/api/getQueryResults/index.ts
@@ -1,0 +1,2 @@
+export { default as getQueryResults } from './getQueryResults';
+export type { getQueryResultsProps } from './types';
