+++ graphql/src/datasources/es/api/getQueryResults/types.ts
@@ -1,0 +1,30 @@
+import type { ElasticLanguage } from '../../../../types';
+import type {
+  AccessibilityProfileType,
+  ElasticSearchIndex,
+  OrderByDistanceParams,
+  OrderByNameParams,
+} from '../../types';
+
+export type OrderByFields = {
+  orderByDistance?: OrderByDistanceParams;
+  orderByName?: OrderByNameParams;
+  orderByAccessibilityProfile?: AccessibilityProfileType;
+};
+
+export type getQueryResultsProps = {
+  text?: string;
+  ontology?: string;
+  administrativeDivisionIds?: string[];
+  ontologyTreeIdOrSets?: string[][];
+  ontologyWordIdOrSets?: string[][];
+  providerTypes?: string[];
+  serviceOwnerTypes?: string[];
+  targetGroups?: string[];
+  mustHaveReservableResource?: boolean;
+  index?: ElasticSearchIndex;
+  from?: number;
+  size?: number;
+  languages?: ElasticLanguage[];
+  openAt?: string;
+} & OrderByFields;
