+++ graphql/src/datasources/es/api/getQueryResults/utils/index.ts
@@ -1,0 +1,4 @@
+export { getOntologyFields } from './getOntologyFields';
+export { getDefaultQuery } from './getDefaultQuery';
+export { getFilters } from './getFilters';
+export { sortQuery } from './sortQuery';
