+++ graphql/src/datasources/es/api/index.ts
@@ -1,7 +1,7 @@
 export { default as getAdministrativeDivisions } from './getAdministrativeDivisions';
 export { default as getOntologyTree } from './getOntologyTree';
 export { default as getOntologyWords } from './getOntologyWords';
-export { default as getQueryResults } from './getQueryResults';
-export type { getQueryResultsProps } from './getQueryResults';
+export { default as getQueryResults } from './getQueryResults/getQueryResults';
+export type { getQueryResultsProps } from './getQueryResults/types';
 export { default as getSuggestions } from './getSuggestions';
 export type { getSuggestionProps } from './getSuggestions';
